Handle Registration error codes (4xx,5xx,6xx)
On receiving a 4xx, 5xx, 6xx response to the REGISTER request, whereby the response contains a Retry-After header field, the UE shall not automatically attempt an initial registration via the same IP-CAN and the same P-CSCF for the amount of time
indicated in the Retry-After header field. If the UE is power cycled, the UE can attempt an initial registration.
If no initial registration occurs within the time period indicated by the Retry-After header field, the counter of unsuccessful initial registration attempts is reset.

Reference: RFC5626 "Managing Client-initiated connections in the Session Initiation Protocol (SIP) IETF RFC".