CR020 Add common history
The history log provider is a virtual content provider in that
it does not store any data itself but is allows for a client
to make queries dynamically combining entries from several
other specified providers per query returning a merged cursor
containing all entries that match the selection query in those
specified providers. Note that only read operations are
supported.

Change-Id: Ibca826b06564e29297e5d4021388965a0faa99a8